chore: move hooks to utils again
