+++ apps/wallet/src/screens/Authentication/Invitation/index.tsx
@@ -12,11 +12,11 @@ import {
 	type ViewStyle,
 } from 'react-native';
 import { Anchor, Button, Input, Text, View } from '@walless/gui';
-import { useSafeAreaInsets } from 'hooks';
 import { showError } from 'modals/Error';
 import { appState } from 'state/app';
 import assets from 'utils/assets';
 import { validateInvitationCode } from 'utils/auth';
+import { useSafeAreaInsets } from 'utils/hooks';
 import { navigate } from 'utils/navigation';
 import { hideNativeKeyboard } from 'utils/system';
