+++ apps/wallet/src/screens/Dashboard/History/index.tsx
@@ -3,8 +3,8 @@ import type { ViewStyle } from 'react-native';
 import type { StackScreenProps } from '@react-navigation/stack';
 import { View } from '@walless/gui';
 import HistoryFeature from 'features/History';
-import { useSafeAreaInsets } from 'hooks';
 import { tabBarHeight } from 'utils/constants';
+import { useSafeAreaInsets } from 'utils/hooks';
 import { type HomeParamList } from 'utils/navigation';
 
 type Props = StackScreenProps<HomeParamList, 'History'>;
