+++ apps/wallet/src/stacks/Explorer/Sidebar.tsx
@@ -3,10 +3,10 @@ import { useEffect } from 'react';
 import type { WidgetDocument } from '@walless/store';
 import type { DrawerContentComponentProps } from 'components/DrawerNavigation';
 import { useDrawerStatus } from 'components/DrawerNavigation';
-import { useSnapshot, useWidgets } from 'hooks';
 import { appState } from 'state/app';
 import { runtimeActions } from 'state/runtime';
 import { widgetActions } from 'state/widget';
+import { useSnapshot, useWidgets } from 'utils/hooks';
 import { navigate } from 'utils/navigation';
 
 import WidgetNavigator from './WidgetNavigator';
