+++ apps/wallet/src/utils/hooks/app.ts
