+++ apps/wallet/src/utils/hooks/index.ts
