+++ apps/wallet/src/utils/hooks/widget.ts
